requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
pgvector
httpx
pydantic
python-docx
python-pptx
pdfminer.six
openai
markdown-it-py
mdurl